delete old intent_parser_server.py and replaced with intent_parser_server2.py
